Add tests for students & courses, modify id property and add postman collection

+++ courses/src/main/java/cl/test/courses/entity/Course.java
@@ -7,24 +7,24 @@ public class Course {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String nameCourse;
     private String teacherCourse;
 
     public Course() {
     }
 
-    public Course(Long id, String nameCourse, String teacherCourse) {
+    public Course(int id, String nameCourse, String teacherCourse) {
         this.id = id;
         this.nameCourse = nameCourse;
         this.teacherCourse = teacherCourse;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
