+++ courses/src/main/java/cl/test/courses/repository/CoursesRepository.java
@@ -3,6 +3,6 @@ package cl.test.courses.repository;
 import cl.test.courses.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoursesRepository extends JpaRepository<Course, Long> {
+public interface CoursesRepository extends JpaRepository<Course, Integer> {
 
 }
