+++ students/src/main/java/cl/test/students/repository/StudentsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StudentsRepository extends JpaRepository<Student, Long> {
-    List<Student> findByCourseId(Long courseId);
+public interface StudentsRepository extends JpaRepository<Student, Integer> {
+    List<Student> findByCourseId(int courseId);
 }
